Neue Kamera-Implementierung und Code-Optimierungen
Diese Änderungen führen eine neue `Camera`-Klasse ein, die eine `UpdatePosition`-Methode zur Aktualisierung der Kameraposition und -rotation basierend auf Benutzereingaben enthält. Die Transformationsstruktur wurde aktualisiert, um die neuen Rotationsparameter `yaw`, `pitch` und `roll` zu integrieren.

Die Methode `UpdateConstantBuffer` in `Graphics.cpp` wurde angepasst, um ein `Camera`-Objekt zu verwenden, was die Handhabung der Kameraposition vereinfacht. In `Scene.cpp` und `MessageEvaluation.cpp` wurden die Logik zur Verarbeitung von Tastenanschlägen und die Frame-Aktualisierung entsprechend aktualisiert.

Zusätzlich wurden die Header-Dateien in `Graphics.h` neu angeordnet, um die Abhängigkeiten zu optimieren, und die Projektdateien wurden aktualisiert, um die neuen Quelldateien einzuschließen. Typanpassungen in `objfileimport.cpp` stellen sicher, dass Berechnungen korrekt durchgeführt werden.

+++ coolerprojektname/objfileimport.cpp
@@ -128,16 +128,16 @@ float objfileimport::findnumberfloataccurate(std::string& line, size_t& size)
 
 	for (size_t i = signbit; i < decimalpoint; i++)
 	{
-		beforecommadigits = beforecommadigits + (unsigned char)(line[i] - 48) * std::pow(10, i - signbit);
+		beforecommadigits = beforecommadigits + (unsigned int)((unsigned char)(line[i] - 48) * std::pow(10, i - signbit));
 		///bilden der Vorkommastellen "std::pow(10, i-signbit)"
 		///steht für 10^(aktuelle 10er Potenz) -> durch addieren 
 		///des Signbit im Iterator, hier entfernen
 	}
 
-	float returnfloat = beforecommadigits; //erstellen der return Float mit den Vorkommastellen
-	int exp = decimalpoint - numberlength; //bilden des Exponent -> "negative Nummernlänge + decimal Punkt" entspricht "decimalpunkt - numernlänge"
+	float returnfloat = (float)beforecommadigits; //erstellen der return Float mit den Vorkommastellen
+	int exp = (int)decimalpoint - (int)numberlength; //bilden des Exponent -> "negative Nummernlänge + decimal Punkt" entspricht "decimalpunkt - numernlänge"
 
-	returnfloat = returnfloat + std::pow(10.0, (double)(exp)) * afterkommadigits; //addieren der Nachkommazahlen -> Kommazahl wird gebildet
+	returnfloat = returnfloat + (float)(std::pow(10.0, (double)(exp)) * afterkommadigits); //addieren der Nachkommazahlen -> Kommazahl wird gebildet
 
 	size = line.length()+1; //returnen der Länge
